vhost: Use a tree to store memory mappings
At the moment, the tree is only used to store 1:1 maps of the qemu
virtual addresses of shadow virtqueue vring and the guest's addresses.
In other words, the tree only serves to check if the address the guest
exposed is valid at the moment qemu receives the miss.

It does not work if device has restrictions in its iova
range at the moment.

Updates to tree are protected by BQL, each one always run from main
event loop context. vhost_device_iotlb_miss runs in the same one on
reading it.

Signed-off-by: Eugenio Pérez <eperezma@redhat.com>