[TODO] vdpa: Add custom IOTLB translations to SVQ
Use translations added in VhostIOVATree in SVQ.

Now every element needs to remember the iova / GPA address also, so
VirtQueue can consume the elements properly. This adds a little
overhead per VQ element, having to allocate more memory to stash them.

As a possible optimization, this allocation could be avoided if the
descriptor is not a chain but a single one, but this is left undone.

TODO: Do we need to store address in SVQElement?

Signed-off-by: Eugenio Pérez <eperezma@redhat.com>