vhost: Add iova_rev_maps_alloc
This way the tree can store/map a completely new virtual address space,
in the range the device admits.

It does not limit the range it will allocate, but the IOVA address of
the maps are allocated growing. A range limitation will be need for the
cases where start_addr != 0.

Tools for remove mappings will be needed also.

Signed-off-by: Eugenio Pérez <eperezma@redhat.com>