vhost-vdpa: never map with vDPA listener
This commit is a workaround that will not go to the final version.

vp_vdpa is not able to reset all IOTLBs, so we force to not to map them
in the first place.

Checkpath detects a few errors because of #if 0 / #endif pairs, but it's
the less intrusive way to comment out all the code we want to skip.
Since this commit is not intended to go to the final series, I left it
that way.

Signed-off-by: Eugenio Pérez <eperezma@redhat.com>