vhost: delay set_vring_ready after DRIVER_OK
To restore the device in the destination of a live migration we send the
commands through control virtqueue. For a device to read CVQ it must
have received DRIVER_OK status bit.

However this open a window where the device could start receiving
packets in rx queue 0 before it receive the RSS configuration. To avoid
that, we will not send vring_enable until all configuration is used by
the device.

As a first step, run vhost_set_vring_ready for all vhost_net backend after
all of them has been started (with DRIVER_OK). This code should not
affect vdpa.

Signed-off-by: Eugenio Pérez <eperezma@redhat.com>