vdpa net: allow VHOST_F_LOG_ALL
Since some actions move to start function instead of init, the device
features may not be the parent vdpa device's, but the one returned by
vhost backend. If transition to SVQ is supported, the vhost backend will
return _F_LOG_ALL to signal the device is migratable.

Add VHOST_F_LOG_ALL. HW dirty page tracking can be added on top of this
change if the device supports it in the future.

Signed-off-by: Eugenio Pérez <eperezma@redhat.com>